+++ http/src/main/scala/zealot/http/http.scala
@@ -79,6 +79,7 @@ trait HttpResponse {
   def document(using session: HttpSession)      : ZLT[HtmlElement]
   def header(name: String)                      : Option[String]
   def requestedUrl                              : String
+  def follow                                    : Boolean = code >= 300 && code < 400
 }
 
 @FunctionalInterface
@@ -628,7 +629,7 @@ case class DefaultHttpRequest (
 
     def handleRedirect(response: HttpResponse): ZLT[HttpResponse] = {
 
-      def follow: ZLT[HttpResponse] = {
+      def follow(toFollow: HttpResponse): ZLT[HttpResponse] = {
 
         def fixRelativeUrl(location: String): ZLT[String] = {
 
@@ -645,19 +646,18 @@ case class DefaultHttpRequest (
         }
 
         for {
-          location <- response.redirect
+          location <- toFollow.redirect
           loc      <- fixRelativeUrl(location)
           req      <- session.requestGiven(loc, version)
           res      <- req.named(s"FR-${name.getOrElse("_")}").get()
         } yield res
       }
 
-      if(request.followRedirects && response.code >= 300 && response.code < 400)
-        for {
-          _      <- interceptor.onFollow(this, response)
-          result <- follow
-        } yield result
-
+      if(request.followRedirects && response.follow)
+        for
+          alternative <- interceptor.onFollow(this, response)
+          result      <- if (alternative.follow) follow(alternative) else ZIO.succeed(alternative)
+        yield result
       else ZIO.succeed(response)
     }