+++ changes.md
@@ -1,5 +1,7 @@
 # Zealot
 
+ - Using the result of `HttpInterceptor.onFollow` when performing redirect requests 
+
 ## Release v0.2.9
 LTS: 25/04/2025